# Vite: Cleanup and prebundle dependencies