Works on https://github.com/storybookjs/storybook/issues/29038

## What I did

|  | **Before** | **After** |
|---|---|---|
| pkg-size | ![image](https://github.com/user-attachments/assets/78f29a01-c2b2-4b69-a286-c3ea3e2d8417) | ![image](https://github.com/user-attachments/assets/f9c1053b-f2dd-4269-8470-5a1e68540135) |
| npmgraph | ![image](https://github.com/user-attachments/assets/dd52ecc9-2fd3-4a10-8b78-562239214cbe) | ![image](https://github.com/user-attachments/assets/9ea454ef-b74a-4190-a973-33a996eafe57) |

### Sources

https://pkg-size.dev/@storybook%2Fbuilder-vite@8.4.0-alpha.5?no-peers

https://pkg-size.dev/@storybook%2Fbuilder-vite@0.0.0-pr-29302-sha-21f4ff11?no-peers

https://npmgraph.js.org/?q=@storybook/builder-vite@8.4.0-alpha.5#hide=

https://npmgraph.js.org/?q=@storybook/builder-vite@0.0.0-pr-29302-sha-21f4ff11#zoom=w


## Checklist for Contributors

### Testing

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to communicate how to test your changes -->

#### The changes in this PR are covered in the following automated tests:
- [x] stories
- [ ] unit tests
- [ ] integration tests
- [x] end-to-end tests

#### Manual testing

_This section is mandatory for all contributions. If you believe no manual test is necessary, please state so explicitly. Thanks!_

<!-- Please include the steps to test your changes here. For example:

1. Run a sandbox for template, e.g. `yarn task --task sandbox --start-from auto --template react-vite/default-ts`
2. Open Storybook in your browser
3. Access X story

-->

### Documentation

<!-- Please check (put an "x" inside the "[ ]") the applicable items below to indicate which documentation has been updated. -->

- [ ] Add or update documentation reflecting your changes
- [ ] If you are deprecating/removing a feature, make sure to update
      [MIGRATION.MD](https://github.com/storybookjs/storybook/blob/next/MIGRATION.md)

## Checklist for Maintainers

- [ ] When this PR is ready for testing, make sure to add `ci:normal`, `ci:merged` or `ci:daily` GH label to it to run a specific set of sandboxes. The particular set of sandboxes can be found in `code/lib/cli/src/sandbox-templates.ts`
- [ ] Make sure this PR contains **one** of the labels below:
   <details>
     <summary>Available labels</summary>

     - `bug`: Internal changes that fixes incorrect behavior.
     - `maintenance`: User-facing maintenance tasks.
     - `dependencies`: Upgrading (sometimes downgrading) dependencies.
     - `build`: Internal-facing build tooling & test updates. Will not show up in release changelog.
     - `cleanup`: Minor cleanup style change. Will not show up in release changelog.
     - `documentation`: Documentation **only** changes. Will not show up in release changelog.
     - `feature request`: Introducing a new feature.
     - `BREAKING CHANGE`: Changes that break compatibility in some way with current major version.
     - `other`: Changes that don't fit in the above categories.
   
   </details>

### 🦋 Canary release

<!-- CANARY_RELEASE_SECTION -->
This pull request has been released as version `0.0.0-pr-29302-sha-21f4ff11`. Try it out in a new sandbox by running `npx storybook@0.0.0-pr-29302-sha-21f4ff11 sandbox` or in an existing project with `npx storybook@0.0.0-pr-29302-sha-21f4ff11 upgrade`.
<details>
<summary>More information</summary>

| | |
| --- | --- |
| **Published version** | [`0.0.0-pr-29302-sha-21f4ff11`](https://npmjs.com/package/storybook/v/0.0.0-pr-29302-sha-21f4ff11) |
| **Triggered by** | @JReinhold |
| **Repository** | [storybookjs/storybook](https://github.com/storybookjs/storybook) |
| **Branch** | [`jeppe/prebundle-builder-vite`](https://github.com/storybookjs/storybook/tree/jeppe/prebundle-builder-vite) |
| **Commit** | [`21f4ff11`](https://github.com/storybookjs/storybook/commit/21f4ff114baef03263d56bf2c8c74e0df4fb74e5) |
| **Datetime** | Tue Oct  8 20:23:17 UTC 2024 (`1728418997`) |
| **Workflow run** | [11243150011](https://github.com/storybookjs/storybook/actions/runs/11243150011) |

To request a new release of this pull request, mention the `@storybookjs/core` team.

_core team members can create a new canary release [here](https://github.com/storybookjs/storybook/actions/workflows/canary-release-pr.yml) or locally with `gh workflow run --repo storybookjs/storybook canary-release-pr.yml --field pr=29302`_
</details>
<!-- CANARY_RELEASE_SECTION -->

<!-- BENCHMARK_SECTION -->

| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createSize |  0 B | 0 B | 0 B | - | - |
| generateSize |  78.7 MB | 78.7 MB | 0 B | 0.83 | 0% |
| initSize |  148 MB | 148 MB | -21.1 kB | **-3.09** | 0% |
| diffSize |  69.3 MB | 69.3 MB | -21.1 kB | **-2.87** | 0% |
| buildSize |  6.79 MB | 6.79 MB | 0 B | **4.22** | 0% |
| buildSbAddonsSize |  1.5 MB | 1.5 MB | 0 B | - | 0% |
| buildSbCommonSize |  195 kB | 195 kB | 0 B | - | 0% |
| buildSbManagerSize |  1.83 MB | 1.83 MB | 0 B | 0.58 | 0% |
| buildSbPreviewSize |  270 kB | 270 kB | 0 B | - | 0% |
| buildStaticSize |  0 B | 0 B | 0 B | - | - |
| buildPrebuildSize |  3.8 MB | 3.8 MB | 0 B | 0.58 | 0% |
| buildPreviewSize |  2.99 MB | 2.99 MB | 0 B | **4.36** | 0% |
| testBuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbAddonsSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbCommonSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbManagerSize |  0 B | 0 B | 0 B | - | - |
| testBuildSbPreviewSize |  0 B | 0 B | 0 B | - | - |
| testBuildStaticSize |  0 B | 0 B | 0 B | - | - |
| testBuildPrebuildSize |  0 B | 0 B | 0 B | - | - |
| testBuildPreviewSize |  0 B | 0 B | 0 B | - | - |



| name | before | after | diff | z | % |
| ---- | ------ | ----- | ---- | - | - |
| createTime |  16.5s | 13.9s | -2s -612ms | 0.05 | -18.7% |
| generateTime |  20.6s | 23.5s | 2.9s | 0.83 | 12.3% |
| initTime |  13.5s | 16.2s | 2.6s | 0.64 | 16.5% |
| buildTime |  8.2s | 9.3s | 1s | -0.36 | 11.3% |
| testBuildTime |  0ms | 0ms | 0ms | - | - |
| devPreviewResponsive |  6.7s | 6.2s | -525ms | -0.89 | -8.5% |
| devManagerResponsive |  4.1s | 4.1s | -33ms | -0.81 | -0.8% |
| devManagerHeaderVisible |  587ms | 508ms | -79ms | -1.18 | -15.6% |
| devManagerIndexVisible |  618ms | 536ms | -82ms | -1.17 | -15.3% |
| devStoryVisibleUncached |  1s | 972ms | -55ms | -0.59 | -5.7% |
| devStoryVisible |  617ms | 535ms | -82ms | -1.18 | -15.3% |
| devAutodocsVisible |  495ms | 455ms | -40ms | -0.87 | -8.8% |
| devMDXVisible |  519ms | 427ms | -92ms | -1.22 | -21.5% |
| buildManagerHeaderVisible |  492ms | 459ms | -33ms | -1.04 | -7.2% |
| buildManagerIndexVisible |  495ms | 460ms | -35ms | -1.15 | -7.6% |
| buildStoryVisible |  525ms | 527ms | 2ms | -0.81 | 0.4% |
| buildAutodocsVisible |  445ms | 426ms | -19ms | -1.06 | -4.5% |
| buildMDXVisible |  496ms | 455ms | -41ms | -0.64 | -9% |

<!-- BENCHMARK_SECTION -->

<!-- greptile_comment -->

## Greptile Summary

This PR optimizes the @storybook/builder-vite package by simplifying dependencies and removing specific configurations.

- Removed several dependencies and moved others to devDependencies in `code/builders/builder-vite/package.json`
- Simplified peerDependencies in `package.json`, potentially reducing package size
- Removed Glimmer-specific plugin configuration from `code/builders/builder-vite/src/vite-config.ts`
- These changes may impact users of the Glimmer framework with Storybook

<!-- /greptile_comment -->