@tobiasdiez 2024-10-10
Did you also check these numbers in "real" applications? Almost every vite / rollup / esbuild plugin uses magic-string if it performs code source manipulation. Other popular usage relevant for storybook are vue and vitest. Thus there is a very high chance that once you use vite, you already have magic-string in your dependency tree. Thus, if you bundle it, then the actual install size increases. Based on the install numbers, a similar argument can probably be made for find-cache-dir.